Feature/139 (#140)
* error: 깃 푸쉬 도중 에러가 나서 comment 클래스 수정

* docs: 패키지 생성

* feat: rq응답 코드 도입

* feat: 신청 확인창 alert로 변경

* refactor: return RsData로 통일.

+++ src/main/java/com/ll/olol/boundedContext/comment/service/CommentService.java
@@ -9,13 +9,12 @@ import com.ll.olol.boundedContext.member.entity.Member;
 import com.ll.olol.boundedContext.notification.event.EventAfterComment;
 import com.ll.olol.boundedContext.recruitment.entity.RecruitmentArticle;
 import com.ll.olol.boundedContext.recruitment.repository.RecruitmentRepository;
+import java.util.List;
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -57,8 +56,12 @@ public class CommentService {
     @Transactional
     public RsData update(Long id, String content) {
         Comment updateComment = findOne(id);
+        RsData rsData = isEqualMemberById(id);
         if (updateComment == null) {
             return RsData.of("F-1", "업데이트 하려는 댓글이 없습니다.");
+        }
+        if (rsData.isFail()) {
+            return isEqualMemberById(id);
         }
         updateComment.setContent(content);
         return RsData.of("S-1", "댓글 수정 성공");
@@ -67,8 +70,12 @@ public class CommentService {
     @Transactional
     public RsData commentDelete(Long id) {
         Optional<Comment> id1 = commentRepository.findById(id);
+        RsData rsData = isEqualMemberById(id);
         if (id1.isEmpty()) {
             return RsData.of("F-1", "삭제하려는 댓글이 없습니다.");
+        }
+        if (rsData.isFail()) {
+            return isEqualMemberById(id);
         }
         commentRepository.delete(id1.get());
         return RsData.of("S-1", "댓글 삭제 성공");
